🎉💣🛠📝 register basclass and refactor existing methods - baseclass is registered and existing methods are integrated - refactoring: fuzzy_scores() name is changed to generic scores, fixes #20 - changed documentation note
